pyproject.toml
[tool.poetry]
name = "wildlife"
version = "0.1.0"
description = "A sample groundlight application"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
groundlight = "^0.14.0"
pillow = "9.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
